+++ src/jpip/query.cc
@@ -1,0 +1,35 @@
+#include "query.h"
+
+#include <cstring>
+
+using namespace std;
+
+namespace jpip {
+
+    Query ParseQuery(const char *begin, const char *end) {
+        Query query;
+
+        while (begin < end) {
+            const char *separator = static_cast<const char *>(memchr(begin, '&', end - begin));
+            const char *parameter_end = separator ? separator : end;
+            const char *equals = static_cast<const char *>(memchr(begin, '=', parameter_end - begin));
+            const char *name_end = equals ? equals : parameter_end;
+
+            query.push_back({string(begin, name_end),
+                             string(equals ? equals + 1 : parameter_end, parameter_end)});
+
+            if (!separator)
+                break;
+            begin = separator + 1;
+        }
+        return query;
+    }
+
+    const string *FindParameter(const Query &query, const char *name) {
+        for (Query::const_reverse_iterator i = query.rbegin(); i != query.rend(); ++i)
+            if (i->name == name)
+                return &i->value;
+        return NULL;
+    }
+
+}